[mac] Make type requests reusable
